Break out source model
Ultimately, I want to try to extract date, time and location for the
meetings from source material.  In order to make this easier, break out
the source URLs as a separate field in the database.

Because other pieces of information might eventually have source URLs,
use Django's Content Type framework to make a source relateable to any
model.

Update the admin and create views to be able to specify source URLs for
meetings.

Add a migration to pull existing sources from the ArchieML.

Remove support for extracting ArchieML fields from the notes field via
the REST API and add an explicit sources field to the response.

Addresses #13
https://github.com/ghing/publicmeetings/issues/13